using internal funktion; only install when DESKTOP_PACKAGES asked for install

+++ template/debootstrap/desktop/15-discord.sh
@@ -1,6 +1,6 @@
-if contains_package "/debootstrap/desktop/15-packages-custom.txt" "discord"; then
+if [ "$INSTALL_DESKTOP_PACKAGES" = "true" ] && contains_package "/debootstrap/desktop/15-packages-custom.txt" "discord"; then
 	log "Installing Discord..."
 	download "https://discordapp.com/api/download?platform=linux&format=deb" /tmp/discord.deb
-	apt install /tmp/discord.deb -y
+	apt_install_package /tmp/discord.deb
 	rm /tmp/discord.deb
 fi
